src/mcp-server/brave_mcp/requirements.txt
boto3
cryptography
fastmcp
mcp[cli]
httpx
pyjwt
python-dotenv
requests
brave-search-python-client 
uvicorn
beautifulsoup4
psutil